+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Reissue::Parser class to parse the changelog file
 - Reissue::Printer class to print the changelog file
 - Reissue::Task class to handle the creation of reissue tasks
+- Return version and date from Reissue.finalize
 
 ### Fixed: